requirements.txt
gunicorn==19.9.0
requests==2.20.0

WTForms==2.1
Flask-WTF==0.14.2
bokeh==0.13.0
flask-bootstrap